+++ application/src/components/index.js
@@ -1,10 +1,21 @@
-import Login from './login/login';
-import Main from './main/main';
-import Nav from './nav/nav';
-import OldOrderForm from './old-order-form/orderForm';
-import OrderFormPage from './order-form-page/order-form-page';
-import Template from './common/template';
-import OldViewOrders from './old-view-orders/viewOrders';
-import ViewOrders from './view-orders/viewOrders';
+import Login from "./login/login";
+import Main from "./main/main";
+import Nav from "./nav/nav";
+import OldOrderForm from "./old-order-form/orderForm";
+import OrderFormPage from "./order-form-page/order-form-page";
+import Template from "./common/template";
+import OldViewOrders from "./old-view-orders/viewOrders";
+import ViewOrders from "./view-orders/viewOrders";
+import OrderForm from "./order-form/order-form";
 
-export { Login, OldOrderForm, OrderFormPage, Main, Nav, Template, OldViewOrders, ViewOrders };
+export {
+  Login,
+  OldOrderForm,
+  OrderFormPage,
+  Main,
+  Nav,
+  Template,
+  OldViewOrders,
+  ViewOrders,
+  OrderForm,
+};
